refactor(print-format-builder): add useColorControl composable
